Simplify TextProvider, Fix all @text ending with space
This commit simplifies TextProvider to be a lot clearer and also avoid swallowing an unnecessary exception.

Furthermore, it changes its behaviour so that created text does not always have a space at the end. Usually, people don't mean to add an unnecessary space at the end of their text argument if they don't put more text.

The motivation for this commit is that I am currently experiencing the space problem in my usage of this library.

Thanks!