Fix only single parameter classifier allowed per parameter class
Before, it was not possible to register more than one classifier
annotation (with a provider) for a single parameter type.

The cause for this was that Key#compareTo(Key) returned 0 (equal) for
any two Key objects that had *any* classifier. Therefore, TreeSet
believed that those were the same, and refused to add another Key
with a different qualifier.

The use case for the changed behaviour is that, as a developer, I
want to be able to register different providers for a single
parameter type. The classifier API suggests that this is possible,
while it actually was not before this commit. Specifically, I tried
to register a new String provider with custom behaviour, classified
by a classifier annotation, but noticed that only the preconfigured
@Text was recognised and mapped to the correct provider. During a
debug session, I found the cause to be what is described above.

The implementation of the fix is quite simple: The compareTo(Key)
method now returns the result of String#compareTo(String) of the
classifier class name if both keys have a classifier. This allows any
Set to accept Key instances with non-null, but different classifiers.
To have compareTo(Key) consistent with equals(Object), I also added
custom #equals(Object) and #hashCode() implementations. It is
also generally considered good practice to implement #equals(Object)
and #hashCode() if the class is to be used as Collection key.